Move token logic to separate RH class
