+++ indico/web/rh.py
@@ -33,6 +33,7 @@ from indico.util.i18n import _
 from indico.util.locators import get_locator
 from indico.util.signals import values_from_signal
 from indico.web.flask.util import url_for
+from indico.web.util import is_signed_url_valid
 
 
 HTTP_VERBS = {'GET', 'PATCH', 'POST', 'PUT', 'DELETE'}
@@ -332,3 +333,15 @@ class RHProtected(RH):
 
     def _check_access(self):
         self._require_user()
+
+
+class RHTokenProtected(RH):
+    """A request handler which is protected through a signature token parameter."""
+
+    def _process_args(self):
+        self.user = db.m.User.get_one(request.view_args['user_id'])
+
+    def _check_access(self):
+        token = request.args.get('token')
+        if not token or not is_signed_url_valid(self.user, request.full_path):
+            raise Forbidden
